requirements.txt
Django~=5.1.6
djoser~=2.3.1
djangorestframework~=3.15.2
requests~=2.32.3